+++ changelog.md
@@ -41,6 +41,7 @@
 * Added Spanish and Italian languages.
 * Added undo/redo.
 * Added password protection for publicly shared grids and forms.
+* Made views trashable.
 
 ## Released (2022-03-03 1.9.1)